♻️ NavComponents, Logo 컴포넌트 분리, 타입 지정 레거시 대신 ReactNode로 수정 (#62)
