+++ src/components/common/Logo.tsx
@@ -1,0 +1,12 @@
+export const Logo = () => {
+  return (
+    <div className="flex h-12 w-36 items-center">
+      <div className="bg-primary-500 text-4 flex h-8 w-8 items-center justify-center rounded-md font-bold text-white">
+        S
+      </div>
+      <div className="text-primary-600 text-heading4 bg-white pl-2 font-bold">
+        StudyHub
+      </div>
+    </div>
+  )
+}
