+++ src/components/index.tsx
@@ -4,6 +4,17 @@ import Button from '@/components/common/Button'
 import ImageCard from '@/components/common/Card/ImageCard'
 import Input from '@/components/common/Input'
 import MyPageSideBar from '@/components/my-page-side-bar/MyPageSideBar'
-import Header from './header/Header'
+import Header from '@/components/header/Header'
+import { Logo } from '@/components/common/Logo'
+import { NavComponent } from './header/NavComponent'
 
-export { Badge, Button, ImageCard, Input, MyPageSideBar, Header }
+export {
+  Badge,
+  Button,
+  ImageCard,
+  Input,
+  MyPageSideBar,
+  Header,
+  Logo,
+  NavComponent,
+}
